Fixing Sitecore.Mvc.Exceptions.CyclicRenderingException.
Sitecore uses the ToString method on the Renderer class to try and detect cyclic renderings.  Since MiniProfilerRenderer wraps the implementing renderer, Sitecore is detecting a cyclic rendering scenario.  This is currently breaking the Forms builder in Sitecore.

To fix, I've overriden the ToString method on MiniProfilerRenderer and returning the ToString value of the inner renderer.